fix: 계산 함수 분리
